Allow multiple prefix keys.
Because multiple prefix keys may use the same suffix key, I cannot bind
the suffix key when establishing the keybinding, for that would result
in the same suffix key being bound twice.

So now when adding suffixes I just remember what needs to be added.  And
then, when the prefix key is hit, I bind all the suffix keys right then
and there.  And once after a suffix key has been hit, I remove the
bindings again that were established when the prefix key was hit.